+++ package.json
@@ -34,6 +34,6 @@
     "react-loading-skeleton": "^3.4.0",
     "tailwindcss": "3.3.3",
     "typescript": "5.2.2",
-    "viem": "*"
+    "viem": "^2.21.10"
   }
 }
